Closed beta, v0.9-ac

+++ Application/Dictionaries/Delete.cs
@@ -27,10 +27,12 @@ namespace Application.Dictionaries
             public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
             {
                 var dictionary = await _context.Dictionaries.FindAsync(request.Id);
-                
+
                 if (dictionary == null)
-                    throw new RestException(HttpStatusCode.NotFound,
-                        new {dictionary = "Not found"});
+                    throw new RestException(HttpStatusCode.NotFound, ErrorType.DictionaryNotFound);
+
+                if (dictionary.IsMain)
+                    throw new RestException(HttpStatusCode.BadRequest, ErrorType.MainDictionaryDeletion);
 
                 _context.Dictionaries.Remove(dictionary);
 
@@ -38,7 +40,7 @@ namespace Application.Dictionaries
 
                 if (success)
                     return Unit.Value;
-                throw new Exception("Problem saving changes.");
+                throw new RestException(HttpStatusCode.InternalServerError, ErrorType.SavingChangesError);
             }
         }
     }
